+++ src/Extension/TableOfContents/Normalizer/AsIsNormalizerStrategy.php
@@ -13,7 +13,7 @@ namespace League\CommonMark\Extension\TableOfContents\Normalizer;
 
 use League\CommonMark\Block\Element\ListBlock;
 use League\CommonMark\Block\Element\ListItem;
-use League\CommonMark\Extension\TableOfContents\TableOfContents;
+use League\CommonMark\Extension\TableOfContents\Node\TableOfContents;
 
 final class AsIsNormalizerStrategy implements NormalizerStrategyInterface
 {
@@ -63,3 +63,6 @@ final class AsIsNormalizerStrategy implements NormalizerStrategyInterface
         $this->lastListItem = $listItemToAdd;
     }
 }
+
+// Trigger autoload without causing a deprecated error
+\class_exists(TableOfContents::class);