+++ src/Extension/TableOfContents/TableOfContents.php
@@ -12,7 +12,19 @@
 namespace League\CommonMark\Extension\TableOfContents;
 
 use League\CommonMark\Block\Element\ListBlock;
+use League\CommonMark\Extension\TableOfContents\Node\TableOfContents as NewTableOfContents;
 
-final class TableOfContents extends ListBlock
-{
+if (!class_exists(NewTableOfContents::class)) {
+    @trigger_error(sprintf('TableOfContents has moved to a new namespace; use %s instead', NewTableOfContents::class), \E_USER_DEPRECATED);
+}
+
+\class_alias(NewTableOfContents::class, TableOfContents::class);
+
+if (false) {
+    /**
+     * @deprecated This class has moved to the Node sub-namespace; use that instead
+     */
+    final class TableOfContents extends ListBlock
+    {
+    }
 }