+++ src/Extension/TableOfContents/TableOfContentsGeneratorInterface.php
@@ -12,8 +12,12 @@
 namespace League\CommonMark\Extension\TableOfContents;
 
 use League\CommonMark\Block\Element\Document;
+use League\CommonMark\Extension\TableOfContents\Node\TableOfContents;
 
 interface TableOfContentsGeneratorInterface
 {
     public function generate(Document $document): ?TableOfContents;
 }
+
+// Trigger autoload without causing a deprecated error
+\class_exists(TableOfContents::class);